tests/integration: Move test utils to one test_utils file
