+++ tests/integration/build_labels/test_build_labels.py
@@ -5,9 +5,9 @@ import json
 import os
 import unittest
 
-from tests.integration.test_podman_compose import podman_compose_path
-from tests.integration.test_podman_compose import test_path
 from tests.integration.test_utils import RunSubprocessMixin
+from tests.integration.test_utils import podman_compose_path
+from tests.integration.test_utils import test_path
 
 
 class TestBuildLabels(unittest.TestCase, RunSubprocessMixin):
