+++ tests/integration/exit-from/test_podman_compose_exit_from.py
@@ -3,9 +3,9 @@
 import os
 import unittest
 
-from tests.integration.test_podman_compose import podman_compose_path
-from tests.integration.test_podman_compose import test_path
 from tests.integration.test_utils import RunSubprocessMixin
+from tests.integration.test_utils import podman_compose_path
+from tests.integration.test_utils import test_path
 
 
 def compose_yaml_path():
