+++ tests/integration/extends_w_file_subdir/test_podman_compose_extends_w_file_subdir.py
@@ -4,9 +4,9 @@ import os
 import unittest
 from pathlib import Path
 
-from tests.integration.test_podman_compose import podman_compose_path
-from tests.integration.test_podman_compose import test_path
 from tests.integration.test_utils import RunSubprocessMixin
+from tests.integration.test_utils import podman_compose_path
+from tests.integration.test_utils import test_path
 
 
 def compose_yaml_path():
