+++ tests/integration/ipam_default/test_podman_compose_ipam_default.py
@@ -4,9 +4,9 @@ import json
 import os
 import unittest
 
-from tests.integration.test_podman_compose import podman_compose_path
-from tests.integration.test_podman_compose import test_path
 from tests.integration.test_utils import RunSubprocessMixin
+from tests.integration.test_utils import podman_compose_path
+from tests.integration.test_utils import test_path
 
 
 def compose_yaml_path():
