+++ tests/integration/lifetime/test_lifetime.py
@@ -6,9 +6,9 @@ import unittest
 
 from parameterized import parameterized
 
-from tests.integration.test_podman_compose import podman_compose_path
-from tests.integration.test_podman_compose import test_path
 from tests.integration.test_utils import RunSubprocessMixin
+from tests.integration.test_utils import podman_compose_path
+from tests.integration.test_utils import test_path
 
 
 class TestLifetime(unittest.TestCase, RunSubprocessMixin):
