+++ tests/integration/nets_test2/test_podman_compose_nets_test2.py
@@ -6,9 +6,9 @@ import unittest
 
 import requests
 
-from tests.integration.test_podman_compose import podman_compose_path
-from tests.integration.test_podman_compose import test_path
 from tests.integration.test_utils import RunSubprocessMixin
+from tests.integration.test_utils import podman_compose_path
+from tests.integration.test_utils import test_path
 
 
 def compose_yaml_path():
