+++ tests/integration/profile/test_podman_compose_up_down.py
@@ -12,9 +12,9 @@ import unittest
 
 from parameterized import parameterized
 
-from tests.integration.test_podman_compose import podman_compose_path
-from tests.integration.test_podman_compose import test_path
 from tests.integration.test_utils import RunSubprocessMixin
+from tests.integration.test_utils import podman_compose_path
+from tests.integration.test_utils import test_path
 
 
 def profile_compose_file():
